TOFIXUP OR DROP: clean/unnification: move "asserts" in the same file
