+++ src/functions/helper_assert.ts
@@ -1,6 +1,32 @@
+import { _t } from "../translation";
 import { Arg, FunctionResultNumber, FunctionResultObject, Matrix, isMatrix } from "../types";
 import { CellErrorType } from "../types/errors";
-import { assert } from "./helpers";
+
+// -----------------------------------------------------------------------------
+// MAIN ASSERTS (ASSERT BY ERROR TYPE)
+// -----------------------------------------------------------------------------
+
+export function assert(condition: boolean, message = _t("Error")): void {
+  if (!condition) {
+    throw { value: CellErrorType.GenericError, message };
+  }
+}
+
+export function assertNotZero(
+  value: number,
+  message = _t("Evaluation of function [[FUNCTION_NAME]] caused a divide by zero error.")
+) {
+  if (value === 0) {
+    throw {
+      value: CellErrorType.DivisionByZero,
+      message,
+    };
+  }
+}
+
+// -----------------------------------------------------------------------------
+// OTHER ASSERTS
+// -----------------------------------------------------------------------------
 
 export function assertSingleColOrRow(errorStr: string, arg: Matrix) {
   assert(arg.length === 1 || arg[0].length === 1, errorStr);
@@ -32,4 +58,14 @@ export function isNumberMatrix(
   arg: Matrix<FunctionResultObject>
 ): arg is Matrix<FunctionResultNumber> {
   return arg.every((row) => row.every((data) => typeof data.value === "number"));
+}
+
+export function assertNumberGreaterThanOrEqualToOne(value: number) {
+  assert(
+    value >= 1,
+    _t(
+      "The function [[FUNCTION_NAME]] expects a number value to be greater than or equal to 1, but receives %s.",
+      value.toString()
+    )
+  );
 }
